requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.6.1
cryptography==43.0.1
redis==5.0.8
imap-tools==1.7.0
pytest==8.3.2
httpx==0.27.2
PyJWT==2.9.0